# [OneDNN] Optimize unsorted segment sum on 1 dim