@cantonios 2024-01-19
> Thank you for the PR and sorry for the delay! Could you please post some performance number comparisons? Are the speedup from avoiding Eigen's indexing overhead?
> 
> @cantonios Would specializing Eigen indexing for 1D case be helpful?

I don't think it's just the indexing - there will be small overhead with `cwiseMax`,  tensor cwise sum, etc.  But yeah, it would be nice to have some numbers to support this.