@AnetaKaczynska 2024-01-22
> > @penpornk @cantonios Here are some performance numbers:
> > Reduction	Num rows	Num cols	Segment size	Before (ns)	After (ns)	Speed up
> > Sum	4,096	1	128	60,812	5,899	10.31x
> > Max	4,096	1	128	52,164	5,524	9.44x
> > Min	4,096	1	128	52,052	5,957	8.74x
> > Prod	4,096	1	128	66,103	9,573	6.91x
> 
> Can you put these numbers up in the PR description so they are added to the commit message?

This PR was submitted by our ex-employee. I don't think I can edit his comment (issue description). Please let me know if there's some other way.